chatbot/chatbot-group4-main/requirements.txt
requests~=2.31.0
textract~=1.6.5
pdfquery~=0.4.3
Scrapy~=2.9.0
beautifulsoup4~=4.8.2
numpy~=1.23.5
pandas~=2.0.2
nltk~=3.8.1
pinecone-client~=2.2.2
tensorflow==2.13.0
nltk~=3.8.1
tensorflow-hub~=0.14.0
bs4~=0.0.1
DateTime~=5.2
PyPDF2~=3.0.1
Flask>=2.3.2
click>=8.1.3
itsdangerous>=2.1.2
Werkzeug>=2.3.3
blinker>=1.6.2
Jinja2>=3.1.2
MarkupSafe>=2.0
Flask-SQLAlchemy>=3.0.5
tiktoken>=0.4.0
scrapy_playwright~=0.0.30